+++ lib/hiptest-publisher/handlebars_helper.rb
@@ -274,5 +274,15 @@ module Hiptest
       binding.pry
       ""
     end
+
+    def hh_if_includes(context, array, element, block_true, block_false = nil)
+      if array&.include?(element)
+        block_true.fn(context)
+      elsif block_false
+        block_false.fn(context)
+      else
+        ''
+      end
+    end
   end
 end