+++ spec/handlebars_helper_spec.rb
@@ -603,4 +603,35 @@ describe Hiptest::HandlebarsHelper do
       expect(evaluate(template, {txt: 'When I plic'})).to eq("When I ploc")
     end
   end
+
+  context 'hh_if_includes' do
+    it 'returns the true block if array contains the element' do
+      template = '{{#if_includes array element}}true block{{else}}false block{{/if_includes}}'
+
+      expect(evaluate(template, array: %w[a b c], element: 'a')).to eq('true block')
+    end
+
+    it 'returns the false block if array does not contain the element' do
+      template = '{{#if_includes array element}}true block{{else}}false block{{/if_includes}}'
+
+      expect(evaluate(template, array: %w[a b c], element: 'd')).to eq('false block')
+    end
+
+    it 'returns the false block if array is empty' do
+      template = '{{#if_includes array element}}true block{{else}}false block{{/if_includes}}'
+
+      expect(evaluate(template, array: %w[], element: 'a')).to eq('false block')
+    end
+
+    it 'returns the false block if array is nil' do
+      template = '{{#if_includes array element}}true block{{else}}false block{{/if_includes}}'
+
+      expect(evaluate(template, array: nil, element: 'a')).to eq('false block')
+    end
+
+    it 'does nothing if there is no false block and element not in array' do
+      template = '{{#if_includes array element}}true block{{/if_includes}}'
+      expect(evaluate(template, array: %w[b c], element: 'a')).to eq('')
+    end
+  end
 end